api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Brendwall Task"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.4.0",
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.9",
    "python-decouple>=3.8",
]
